Rename MKL -> _MKL
Prevent a namespace conflict with the conda mkl service module.
