+++ qutip/_mkl/__init__.py
@@ -1,0 +1,3 @@
+import qutip.settings as qset
+from qutip._mkl.utilities import _set_mkl
+_set_mkl()
