--- qutip/mkl/__init__.py
@@ -1,3 +1,0 @@
-import qutip.settings as qset
-from qutip.mkl.utilities import _set_mkl
-_set_mkl()
